Fix mouse spinning bug in wgl client when the game window is moved left or above the default monitor (negative window position).
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12358 d7cf8633-e32d-0410-b094-e92efae38249

+++ vid_wgl.c
@@ -567,8 +567,8 @@ LONG WINAPI MainWndProc (HWND hWnd, UINT uMsg, WPARAM  wParam, LPARAM lParam)
 			break;
 
 		case WM_MOVE:
-			window_x = (int) LOWORD(lParam);
-			window_y = (int) HIWORD(lParam);
+			window_x = (short) LOWORD(lParam);
+			window_y = (short) HIWORD(lParam);
 			VID_SetMouse(false, false, false);
 			break;
